feat(card): add actions mixin, card templates, and discovery mixin
Added three new modules under src/card/: actions-mixin.js provides a mixin with methods to handle 'More Info' events, custom actions, and service calls with error handling. card-templates.js introduces Lit template functions for rendering card name, status line, and tip buttons, improving UI consistency and localization. discovery-mixin.js adds a mixin for discovering and caching entity information, supporting translation keys, device class, and battery fallback. These changes enable richer card functionality, better status display, and easier integration of new device types.
