+++ src/landroid-card.js
@@ -1,23 +1,21 @@
-import { LitElement, html, nothing } from 'lit';
-import { fireEvent } from 'custom-card-helpers'; // computeStateDisplay,
+import { LitElement, html } from 'lit';
 import registerTemplates from 'ha-template';
 import localize from './localize';
 import styles from './styles';
-import defaultImage from './landroid.svg';
 import { version } from '../package.json';
 import './landroid-card-editor';
-import {
-  isObject,
-  wifiStrengthToQuality,
-  resolveImageMode,
-  resolveHaStateImageTag,
-  getEntityDomain,
-  isSupportedEntity,
-  resolveService
-} from './helpers';
+import { getEntityDomain, isSupportedEntity } from './helpers';
 import * as consts from './constants';
 import { DEFAULT_LANG, defaultConfig } from './defaults';
 import LandroidCardEditor from './landroid-card-editor';
+import { ActionsMixin } from './card/actions-mixin';
+import { DiscoveryMixin } from './card/discovery-mixin';
+import { ImageMixin } from './card/image-mixin';
+import {
+  renderCardName,
+  renderCardStatus,
+  renderTipButton,
+} from './card/card-templates';
 import './elements/lc-linear-progress';
 import './elements/lc-toolbar';
 import './elements/lc-stats';
@@ -50,9 +48,7 @@ console.info(
   'color: #ec6a36; background: white; font-weight: 700; border: 1px #ec6a36 solid; border-radius: 0px 4px 4px 0px;',
 );
 
-class LandroidCard extends LitElement {
-  _huiCardCache = new Map();
-  __patchedCache = new WeakMap();
+class LandroidCard extends ActionsMixin(DiscoveryMixin(ImageMixin(LitElement))) {
 
   /**
    * Properties of the LandroidCard element
@@ -158,91 +154,6 @@ class LandroidCard extends LitElement {
   }
 
   /**
-   * The state of the camera entity specified in the config, or undefined if no camera is specified.
-   *
-   * @return {import('home-assistant').State | undefined} The state of the camera entity specified in the config, or undefined if no camera is specified.
-   */
-  get camera() {
-    return this.hass?.states[this.config.camera];
-  }
-
-  get cameraView() {
-  return this.config?.camera_view ?? defaultConfig.camera_view;
-  }
-
-  get cameraControls() {
-    return this.config?.camera_controls ?? defaultConfig.camera_controls;
-  }
-
-  get cameraMuted() {
-    return this.config?.camera_muted ?? defaultConfig.camera_muted;
-  }
-
-  /**
-   * Returns the URL of the image to display on the card.
-   * If the user has specified an image in the config, that image is returned.
-   * Otherwise, the default image is returned.
-   *
-   * @return {string} The URL of the image to display on the card.
-   */
-  /**
-   * Image rendering mode: 'ha' (standard animated HA element),
-   * 'bundled' (landroid.svg) or 'url' (user image / media-source).
-   *
-   * @return {'ha'|'bundled'|'url'} The resolved image mode.
-   */
-  get imageMode() {
-    return resolveImageMode(this.config?.image);
-  }
-
-  /**
-   * Tag of the standard HA status image element for the entity domain.
-   *
-   * @return {string} 'ha-state-control-vacuum-status' or
-   *   'ha-state-control-lawn_mower-status'.
-   */
-  get haStateImageTag() {
-    return resolveHaStateImageTag(this.config?.entity);
-  }
-
-  get image() {
-    // Бандл-SVG — доступен всегда
-    if (this.imageMode === 'bundled') {
-      return defaultImage;
-    }
-
-    // Пользовательское изображение (URL или резолвленный media-source)
-    if (this._resolvedImage) {
-      return this._resolvedImage;
-    }
-
-    // Запасной вариант, пока media-source резолвится (и для preview в режиме 'ha')
-    return defaultImage;
-  }
-
-  /**
-   * Returns the size of the image in pixels.
-   * The image size is determined by the 'image_size' option in the config,
-   * and defaults to 200 pixels if not specified.
-   *
-   * @return {number} The size of the image in pixels.
-   */
-  get imageSize() {
-    return (this.config?.image_size ?? defaultConfig.image_size) * 50;
-  }
-
-  /**
-   * Returns a CSS style string to position the image on the left or not.
-   * If the user has specified 'image_left: true' in the config, this function returns 'float: left;',
-   * otherwise, an empty string is returned.
-   *
-   * @return {string} A CSS style string to position the image on the left or not.
-   */
-  get imageLeft() {
-    return (this.config?.image_left ?? defaultConfig.image_left) ? 'float: left;' : '';
-  }
-
-  /**
    * Returns whether or not to show the animation on the card.
    * If the user has not specified the 'show_animation' option in the config,
    * this function returns default value.
@@ -312,155 +223,6 @@ class LandroidCard extends LitElement {
    */
   get showToolbar() {
     return this.config?.show_toolbar ?? defaultConfig.show_toolbar;
-  }
-
-  /**
-   * Returns the list of entities to be displayed as settings in the card.
-   * If the user has not specified the 'settings_card' option in the config,
-   * this function returns an empty array.
-   *
-   * @return {string[]} The list of entities to be displayed as settings in the card.
-   */
-  get settingsCardEntities() {
-    if (
-      this.__settingsCache?.hass === this.hass &&
-      this.__settingsCache?.config === this.config
-    ) {
-      return this.__settingsCache.result;
-    }
-
-    const configured = this.config?.settings_card || this.config?.settings;
-    let result = null;
-
-    if (configured?.length) {
-      result = configured;
-    } else {
-      const deviceId = this.hass?.entities?.[this.config.entity]?.device_id;
-      if (deviceId && this.hass?.entities) {
-        const entities = this._deviceEntities
-          .filter(
-            (e) =>
-              e.entity_category === 'config' &&
-              this.hass.states[e.entity_id] &&
-              this.hass.states[e.entity_id].state !== consts.STATE_UNAVAILABLE,
-          )
-          .map((e) => e.entity_id)
-          .sort();
-        result = entities.length ? entities : null;
-      }
-    }
-
-    this.__settingsCache = { hass: this.hass, config: this.config, result };
-    return result;
-  }
-
-  /**
-   * Returns the list of entities to be displayed on the card.
-   * If the user has not specified the 'entities' option in the config,
-   * this function returns an empty object.
-   *
-   * @return {Object} The list of entities to be displayed on the card.
-   */
-  get cardEntities() {
-    if (
-      this.__cardEntitiesCache?.hass === this.hass &&
-      this.__cardEntitiesCache?.config === this.config
-    ) {
-      return this.__cardEntitiesCache.result;
-    }
-
-    const result = Object.fromEntries(
-      Object.entries(consts.CARD_MAP).map(([cardType, card]) => {
-        const configured = this.config?.[cardType + '_card'];
-        let entities;
-        if (configured?.length) {
-          entities = configured.filter((row) => {
-            // поддержка объектных строк и в пользовательском конфиге тоже
-            const eid = typeof row === 'string' ? row : row?.entity;
-            return (
-              this.hass.states[eid] &&
-              this.hass.states[eid].state !== consts.STATE_UNAVAILABLE
-            );
-          });
-        } else {
-          entities = this.findEntitiesByTranslationKeys(card.translationKeys);
-          if (entities.length === 0) {
-            entities = this.findEntitiesByDeviceClass(
-              consts.DEVICE_CLASS_MAP[cardType] ?? [],
-            );
-          }
-        }
-
-        // Батарея vacuum — независимо от автодетекта и ручного конфига
-        if (cardType === consts.BATTERYCARD) {
-          entities = this._withBatteryRow(entities);
-        }
-
-        return [cardType, { entities, labelPosition: card.labelPosition }];
-      }),
-    );
-
-    this.__cardEntitiesCache = { hass: this.hass, config: this.config, result };
-    return result;
-  }
-
-  /**
-   * Добавляет attribute-row батареи в начало списка, если батареи там ещё нет.
-   * Доменно-нейтрально: срабатывает для любой основной сущности с атрибутом
-   * battery_level (vacuum, часть lawn_mower-интеграций вроде Gardena/Husqvarna,
-   * template-конфигурации). Для Landroid Cloud не сработает — там батарея
-   * отдельный сенсор, который найдёт hasBattery.
-   *
-   * @param {Array} entities - Текущий список строк карточки.
-   * @return {Array} Список с battery-row первым элементом или исходный список.
-   */
-  _withBatteryRow(entities) {
-    if (this.entity?.attributes?.battery_level == null) {
-      return entities;
-    }
-
-    const hasBattery = entities.some((row) => {
-      if (row?.type === 'attribute' && row?.attribute === 'battery_level') {
-        return true;
-      }
-      const eid = typeof row === 'string' ? row : row?.entity;
-      return this.hass.states[eid]?.attributes?.device_class === 'battery';
-    });
-    if (hasBattery) return entities;
-
-    return [
-      {
-        type: 'attribute',
-        entity: this.config.entity,
-        attribute: 'battery_level',
-        name: localize('attr.battery') || 'Battery',
-        suffix: '%',
-        icon: 'mdi:battery',
-      },
-      ...entities,
-    ];
-  }
-
-  /**
-   * Returns a list of entities associated with the device of the given lawn_mower entity.
-   * The list is cached for performance reasons.
-   * @return {array} A list of entities associated with the device of the given lawn_mower entity.
-   * @private
-   */
-  get _deviceEntities() {
-    const deviceId = this.hass?.entities?.[this.config.entity]?.device_id;
-    if (!deviceId) return [];
-    if (
-      this.__deviceEntitiesCache?.deviceId === deviceId &&
-      this.__deviceEntitiesCache?.hass === this.hass
-    ) {
-      return this.__deviceEntitiesCache.list;
-    }
-    const list = Object.values(this.hass.entities).filter(
-      (e) => e.device_id === deviceId,
-    );
-    this.__deviceEntitiesCache = { deviceId, hass: this.hass, list };
-    return list;
   }
 
   /**
@@ -625,247 +387,6 @@ class LandroidCard extends LitElement {
   }
 
   /**
-   * Handles the "More Info" action for a given entity.
-   * If the `entityId` is an instance of an Event, it retrieves the `entityId` from the `dataset` of the event's current target.
-   * If the `entityId` is not provided or is `null` or `undefined`, it logs an error message and returns.
-   * Otherwise, it calls the `fireEvent` function with the current component (`this`), the event type `'hass-more-info'`, an object containing the `entityId`, and an options object with `bubbles` set to `false` and `composed` set to `true`.
-   *
-   * @param {string|Event} [entityId=this.entity?.entity_id] - The ID of the entity.
-   * @returns {void}
-   */
-  handleMore(entityId = this.entity?.entity_id) {
-    // Если вызвано из обработчика клика (Lit передаёт Event)
-    if (entityId instanceof Event) {
-      entityId = entityId.currentTarget?.dataset?.entityId
-                ?? this.entity?.entity_id;
-    }
-
-    if (!entityId) {
-      console.error('handleMore: entityId is null or undefined');
-      return;
-    }
-
-    fireEvent(
-      this,
-      'hass-more-info',
-      { entityId },
-      { bubbles: false, composed: true },
-    );
-  }
-
-  _handleCustomEvent(e) {
-    this.handleMore(e.detail?.entityId);
-  }
-
-  /**
-   * Calls a HA service and handles errors gracefully.
-   *
-   * @param {Event} e
-   * @param {string} service - e.g. 'lawn_mower.start_mowing'
-   * @param {Object} serviceData
-   */
-  async callService(e, service, serviceData = {}) {
-    let [domain, name] = service.split('.');
-
-    // Дефолтные сервисы заданы для lawn_mower — ремапим под домен сущности.
-    // Пользовательские сервисы (через actions:/shortcuts:) не затрагиваются,
-    // т.к. идут через callAction, а не сюда.
-    if (
-      domain === consts.LAWNMOWER_SERVICE &&
-      this.entityDomain !== consts.LAWNMOWER_SERVICE
-    ) {
-      name = resolveService(this.entityDomain, name);
-      domain = this.entityDomain;
-    }
-
-    const { isRequest = false, ...service_data } = serviceData;
-
-    try {
-      await this.hass.callService(domain, name, service_data);
-    } catch (err) {
-      console.error(
-        `%c LANDROID-CARD %c ${version} `,
-        'color: white; background: #ec6a36; font-weight: 700;',
-        'color: #ec6a36; background: white; font-weight: 700;',
-        `Service call ${service} failed:`,
-        err,
-      );
-      this.requestInProgress = false;
-      return;
-    }
-
-    if (isRequest) {
-      this.requestInProgress = true;
-      this.requestUpdate();
-    }
-  }
-
-  /**
-   * Calls a service based on the action parameter and the actions config object.
-   *
-   * @param {string} action - The action to call, e.g. `start_mowing`, `edgecut`, `pause`, `dock`.
-   * @param {Object} [params] - An object of options.
-   * @param {string} [params.defaultService] - The service to call if the action is not found in the actions config object.
-   * @return {Function} The function to call to trigger the action.
-   */
-  async handleAction(e, action, params = {}) {
-    try {
-      const actions = this.config.actions || {};
-      const { defaultService = action, ...service_data } = params;
-      delete service_data.action;
-      actions[action]
-        ? await this.callAction(actions[action])
-        : await this.callService(e, defaultService, service_data);
-    } catch (err) {
-      console.error('LANDROID-CARD: handleAction failed:', err);
-      fireEvent(this, 'hass-notification', {
-        message: `Landroid Card: ${err?.message ?? String(err)}`,
-      });
-    }
-  }
-
-  /**
-   * Calls a service based on the action parameter.
-   *
-   * @param {Object} action - An object containing information about the action to call.
-   * @param {string} action.action - The type of action to call, e.g. `perform-action`, `navigate`, `url`, `more-info`.
-   * @param {string} [action.perform_action] - The service to call if `action.action` is `perform-action`.
-   * @param {Object} [action.data] - Data to pass to the service if `action.action` is `perform-action`.
-   * @param {string} [action.target] - Target to pass to the service if `action.action` is `perform-action`.
-   * @param {string} [action.navigation_path] - The path to navigate to if `action.action` is `navigate`.
-   * @param {string} [action.url_path] - The URL to open if `action.action` is `url`.
-   * @param {string} [action.url_target] - The target to open the URL in if `action.action` is `url`.
-   * @param {string} [action.entity] - The entity to open the more info dialog for if `action.action` is `more-info`.
-   * @return {void} This function does not return anything.
-   */
-  async callAction(action) {
-    if (!action?.action) return;
-
-    switch (action.action) {
-      case 'perform-action': {
-        if (!action.perform_action) return;
-        const [domain, service] = action.perform_action.split('.');
-        try {
-          await this.hass.callService(
-            domain,
-            service,
-            action.data ?? {},
-            action.target,
-          );
-        } catch (err) {
-          console.error(
-            `LANDROID-CARD: perform-action ${action.perform_action} failed:`,
-            err,
-          );
-        }
-        break;
-      }
-
-      case 'navigate':
-        window.history.pushState(null, '', action.navigation_path);
-        fireEvent(window, 'location-changed');
-        break;
-
-      case 'url':
-        window.open(action.url_path, action.url_target ?? '_blank');
-        break;
-
-      case 'more-info':
-        this.handleMore(action.entity);
-        break;
-
-      default:
-        console.warn(`LANDROID-CARD: Unknown action type: ${action.action}`);
-    }
-  }
-
-  /**
-   * Retrieves the friendly name of an entity, stripping the device name from it.
-   * If the entity is not found or does not have a friendly name, returns an empty string.
-   *
-   * @param {string} entityId - The entity ID to retrieve the friendly name for.
-   * @return {string} The friendly name of the entity, stripped of the device name if present.
-   */
-  getEntityName(entityId) {
-    const entity = this.hass.states[entityId];
-    if (!isObject(entity)) return '';
-
-    const deviceName = this.entity?.attributes?.friendly_name ?? '';
-    const entityName = entity.attributes?.friendly_name ?? '';
-
-    return entityName.replace(`${deviceName} `, '');
-  }
-
-  /**
-   * Retrieves an entity object from the Home Assistant state object that matches the given translation key.
-   * If no entity is found, returns undefined.
-   * Only returns entities that are not 'unavailable'.
-   *
-   * @param {string} translationKey - The translation key to search for.
-   * @return {object|undefined} The entity object from the Home Assistant state object, or undefined if not found.
-   */
-  getEntityByTranslationKey(translationKey) {
-    const found = this._deviceEntities.find(
-      (e) => e.translation_key === translationKey,
-    );
-    return found ? this.hass.states[found.entity_id] : undefined;
-  }
-
-  /**
-   * Finds entities in the Home Assistant state object that match the given translation keys.
-   * Returns an array of entity IDs that match the given translation keys.
-   * Only returns entities that are not 'unavailable'.
-   * @param {string[]} translationKeys - The translation keys to search for.
-   * @return {string[]} An array of entity IDs that match the given translation keys.
-   */
-  findEntitiesByTranslationKeys(translationKeys) {
-    const deviceEntities = this._deviceEntities;
-    return translationKeys.reduce((result, key) => {
-      const found = deviceEntities.find((e) => e.translation_key === key);
-      if (!found) return result;
-      const stateObj = this.hass.states[found.entity_id];
-      if (stateObj && stateObj.state !== consts.STATE_UNAVAILABLE) {
-        result.push(found.entity_id);
-      }
-      return result;
-    }, []);
-  }
-
-  /**
-   * Finds entities associated with the device that match the given device classes.
-   * Used as a fallback when translation_key lookup returns no results —
-   * typically for non-Landroid integrations (e.g. Mammotion, Husqvarna Automower).
-   *
-   * The device_class is resolved in the following priority order:
-   * 1. `state_obj.attributes.device_class` (runtime state)
-   * 2. `entity_registry.device_class` (user override)
-   * 3. `entity_registry.original_device_class` (integration default)
-   *
-   * @param {string[]} deviceClasses - List of device class strings to match against,
-   *   e.g. `['battery', 'voltage']`. Returns an empty array if the list is empty.
-   * @return {string[]} Array of entity IDs whose device_class matches one of the
-   *   provided values. Excludes unavailable entities.
-   */
-  findEntitiesByDeviceClass(deviceClasses) {
-    if (!deviceClasses.length) return [];
-    const deviceEntities = this._deviceEntities;
-    return deviceEntities.reduce((result, e) => {
-      const stateObj = this.hass.states[e.entity_id];
-      if (!stateObj || stateObj.state === consts.STATE_UNAVAILABLE) return result;
-
-      const dc =
-        stateObj.attributes.device_class ??
-        e.device_class ??
-        e.original_device_class;
-
-      if (dc && deviceClasses.includes(dc)) {
-        result.push(e.entity_id);
-      }
-      return result;
-    }, []);
-  }
-
-  /**
    * Toggles the visibility of the given card type and updates the component to reflect the change.
    *
    * @param {Event} e - The event object.
@@ -880,42 +401,6 @@ class LandroidCard extends LitElement {
   }
 
   /**
-   * Returns a patched version of the state object for the given entity ID.
-   * If the state object has not changed since the last call, returns the cached patched version.
-   * Otherwise, returns a new patched version with the icon and device class attributes set to the values from the registry object if available.
-   * @param {string} entityId - The entity ID to retrieve the patched state object for.
-   * @return {Object|undefined} The patched state object, or undefined if the entity ID is not found in the Home Assistant state object.
-   */
-  getPatchedStateObj(entityId) {
-    const stateObj = this.hass.states[entityId];
-    if (!stateObj) return undefined;
-
-    // Если этот конкретный объект состояния уже пропатчен — возвращаем кэш
-    if (this.__patchedCache.has(stateObj)) {
-      return this.__patchedCache.get(stateObj);
-    }
-
-    const registryObj = this.hass.entities[entityId];
-    const patched = {
-      ...stateObj,
-      attributes: {
-        ...stateObj.attributes,
-        icon:
-          stateObj.attributes.icon ??
-          registryObj?.icon ??
-          registryObj?.original_icon,
-        device_class:
-          stateObj.attributes.device_class ??
-          registryObj?.device_class ??
-          registryObj?.original_device_class,
-      },
-    };
-
-    this.__patchedCache.set(stateObj, patched);
-    return patched;
-  }
-
-  /**
    * Renders a tip button for a given card.
    * label = 0; // none: 0, left: 1 or right: 2
    *
@@ -923,246 +408,7 @@ class LandroidCard extends LitElement {
    * @return {TemplateResult|nothing} The rendered tip button or nothing if the card type is not valid.
    */
   renderTipButton(cardType) {
-    const card = this.cardEntities[cardType];
-    if (!card) {
-      return nothing;
-    }
-
-    // entities может содержать объектные строки (type: 'attribute'
-    // для battery fallback у vacuum) — достаём entity_id в обоих случаях
-    const first = card.entities?.[0];
-    const entityId = typeof first === 'string' ? first : first?.entity;
-    if (!entityId) {
-      return nothing;
-    }
-
-    const entity = this.getPatchedStateObj(entityId);
-    if (!entity) {
-      return nothing;
-    }
-
-    const title = this.getEntityName(entityId);
-
-    // Для attribute-строки показываем значение атрибута, а не состояние
-    const isAttrRow = typeof first !== 'string' && first?.type === 'attribute';
-    const attrValue = isAttrRow ? entity.attributes?.[first.attribute] : null;
-
-    const translationKey = this.hass.entities?.[entityId]?.translation_key;
-    const state =
-      translationKey === consts.TK_SENSOR_WIFI
-        ? wifiStrengthToQuality(entity.state)
-        : isAttrRow && attrValue != null
-          ? `${attrValue}${first.suffix ?? ''}`
-          : this.hass.formatEntityState(entity);
-
-    const labelContent = html`<div .title="${title}: ${state}">${state}</div>`;
-
-    return html`
-      <div class="tip"
-        @click=${this._toggleCardVisibility}
-        data-card-type=${cardType}
-      >
-        ${card.labelPosition === 1 ? labelContent : ''}
-        <state-badge
-          .hass=${this.hass}
-          .stateObj=${entity}
-          .title=${`${title}: ${state}`}
-          .stateColor=${true}
-        ></state-badge>
-        ${card.labelPosition === 2 ? labelContent : ''}
-      </div>
-    `;
-  }
-
-  /**
-   * Renders the camera or image based on the provided state.
-   *
-   * @param {string} state - The state used as a CSS class.
-   * @return {TemplateResult|nothing} The rendered camera or image as a lit-html TemplateResult or nothing.
-   */
-  renderCameraOrImage(state) {
-    if (this.compactView) {
-      return nothing;
-    }
-
-    const cameraEntity = this.hass?.states[this.config.camera];
-
-    if (cameraEntity) {
-      return html`
-        <ha-camera-stream
-          style="height: ${this.imageSize}px; ${this.imageLeft}"
-          class="camera"
-          .hass=${this.hass}
-          .stateObj=${cameraEntity}
-          .cameraView=${this.cameraView}
-          .controls=${this.cameraControls}
-          .muted=${this.cameraMuted}
-          @click=${this.handleMore} data-entity-id=${this.config.camera}
-        ></ha-camera-stream>
-      `;
-    }
-
-    if (this.imageMode === 'ha') {
-      return this.renderHaStateImage();
-    }
-
-    if (this.image) {
-      return html`
-        <div class="landroid-wrapper ${this.showAnimation ? state : ''}">
-          <img
-            style="height: ${this.imageSize}px; ${this.imageLeft}"
-            class="landroid ${this.showAnimation ? state : ''}"
-            src="${this.image}"
-            @click=${this.handleMore}
-          />
-        </div>
-      `;
-    }
-
-    return nothing;
-  }
-
-  /**
-   * Renders the standard animated Home Assistant status image
-   * (`ha-state-control-vacuum-status` / `ha-state-control-lawn_mower-status`).
-   * Falls back to the bundled SVG while the element is not registered
-   * (HA < 2026.5 or the preload is still in flight).
-   *
-   * @return {TemplateResult|nothing} The rendered HA state image.
-   */
-  renderHaStateImage() {
-    const tag = this.haStateImageTag;
-    if (!this.entity) {
-      return nothing;
-    }
-
-    if (!customElements.get(tag)) {
-      // Элемент регистрируется HA лениво — запускаем предзагрузку (идемпотентно)
-      // и показываем бандл-SVG, пока он не появится
-      this._preloadHaStateImage();
-      return this.renderBundledImage('docked');
-    }
-
-    // Элемент HA имеет фиксированный размер 200×200 внутри своего shadow DOM,
-    // поэтому масштабируем его трансформом под настроенный размер изображения
-    const scale = this.imageSize / 200;
-    return html`
-      <div
-        class="ha-state-image"
-        style="height: ${this.imageSize}px; ${this.imageLeft}"
-        @click=${this.handleMore}
-      >
-        <div class="ha-state-image-scale" style="transform: scale(${scale});">
-          ${tag === consts.HA_STATE_IMAGE_TAGS.vacuum
-            ? html`
-                <ha-state-control-vacuum-status
-                  .stateObj=${this.entity}
-                ></ha-state-control-vacuum-status>
-              `
-            : html`
-                <ha-state-control-lawn_mower-status
-                  .stateObj=${this.entity}
-                ></ha-state-control-lawn_mower-status>
-              `}
-        </div>
-      </div>
-    `;
-  }
-
-  /**
-   * Renders the bundled landroid.svg (fallback and `image: 'default'`).
-   *
-   * @param {string} state - The state used as a CSS class.
-   * @return {TemplateResult} The rendered bundled image.
-   */
-  renderBundledImage(state) {
-    return html`
-      <div class="landroid-wrapper ${this.showAnimation ? state : ''}">
-        <img
-          style="height: ${this.imageSize}px; ${this.imageLeft}"
-          class="landroid ${this.showAnimation ? state : ''}"
-          src="${defaultImage}"
-          @click=${this.handleMore}
-        />
-      </div>
-    `;
-  }
-
-  /**
-   * Preloads the HA status image element without opening the more-info dialog.
-   * `window.loadCardHelpers()` is the official runtime helper for custom
-   * cards; `importMoreInfoControl(domain)` triggers HA's own lazy chunk import
-   * which registers the status element (no UI side effects).
-   *
-   * @return {void}
-   */
-  async _preloadHaStateImage() {
-    const tag = this.haStateImageTag;
-    if (customElements.get(tag)) {
-      return;
-    }
-
-    // Запускаем ленивый импорт чанка HA (идемпотентно). Если элемент так и не
-    // зарегистрировался (например, loadCardHelpers ещё не появился на момент
-    // первого рендера), через 2 с снимаем блокировку и даём повторить на
-    // следующем рендере.
-    if (!this.__haImagePreload) {
-      this.__haImagePreload = true;
-      try {
-        const helpers = await window.loadCardHelpers?.();
-        helpers?.importMoreInfoControl?.(
-          tag === consts.HA_STATE_IMAGE_TAGS.vacuum ? 'vacuum' : 'lawn_mower',
-        );
-      } catch (err) {
-        console.warn('LANDROID-CARD: HA state image preload failed', err);
-      } finally {
-        setTimeout(() => {
-          this.__haImagePreload = false;
-        }, 2000);
-      }
-    }
-
-    // Слушаем регистрацию БЕЗ таймаута: когда бы HA ни определил элемент,
-    // перерисовываемся. (Раньше был Promise.race с 2 с — если чанк грузился
-    // дольше, перерисовки не происходило и bundled висел до любого клика.)
-    if (!this.__haImageWatch) {
-      this.__haImageWatch = true;
-      customElements
-        .whenDefined(tag)
-        .then(() => {
-          // Реактивное свойство, а не голый requestUpdate(): shouldUpdate()
-          // отменяет «пустые» обновления, из-за чего bundled висел до клика
-          // по секции (клик меняет _activeCard — он в allowlist)
-          this._haStateImageReady = true;
-        })
-        .catch(() => {});
-    }
-  }
-
-  /**
-   * Applies `show_animation` to the HA status image: pauses/resumes its CSS
-   * animations via the Web Animations API (`getAnimations({ subtree: true })`
-   * covers the element's shadow DOM).
-   *
-   * @return {void}
-   */
-  _syncHaAnimations() {
-    if (this.imageMode !== 'ha') {
-      return;
-    }
-    const host = this.renderRoot?.querySelector('.ha-state-image');
-    if (!host?.getAnimations) {
-      return;
-    }
-    for (const animation of host.getAnimations({ subtree: true })) {
-      if (animation instanceof CSSAnimation) {
-        if (this.showAnimation) {
-          animation.play();
-        } else {
-          animation.pause();
-        }
-      }
-    }
+    return renderTipButton(this, cardType);
   }
 
   /**
@@ -1171,15 +417,7 @@ class LandroidCard extends LitElement {
    * @return {TemplateResult} The rendered name as a lit-html TemplateResult or nothing.
    */
   renderName() {
-    if (!this.showName || !this.entity) return nothing;
-
-    const name = this.entity.attributes?.friendly_name || this.entity.entity_id;
-
-    return html`
-      <div class="landroid-name" title=${name} @click=${this.handleMore}>
-        ${name}
-      </div>
-    `;
+    return renderCardName(this);
   }
 
   /**
@@ -1188,173 +426,14 @@ class LandroidCard extends LitElement {
    * @return {TemplateResult} The rendered status as a lit-html TemplateResult or nothing.
    */
   renderStatus() {
-    if (!this.showStatus) return nothing;
-
-    const state = this.entity?.state;
-
-    // Все опциональные сущности — если нет, просто undefined
-    const zoneSensor = this.getEntityByTranslationKey(consts.TK_SELECT_ZONE);
-    const partyMode = this.getEntityByTranslationKey(consts.TK_SWITCH_PARTY);
-    const lockMode = this.getEntityByTranslationKey(consts.TK_SWITCH_LOCK);
-    const errorSensor = this.getEntityByTranslationKey(consts.TK_SENSOR_ERROR);
-
-    const hasError =
-      isObject(errorSensor) &&
-      errorSensor.state !== 'no_error' &&
-      errorSensor.state !== consts.STATE_UNAVAILABLE;
-
-    let localizedStatus = this.hass.formatEntityState(this.entity) || 'Unknown';
-
-    // rain delay — только если есть сенсор дождя
-    if (state === consts.STATE_RAINDELAY) {
-      const rainSensor = this.getEntityByTranslationKey(
-        consts.TK_SENSOR_RAINDELAY,
-      );
-      if (isObject(rainSensor) && rainSensor.state !== consts.STATE_UNAVAILABLE) {
-        localizedStatus += ` (${this.hass.formatEntityState(rainSensor)})`;
-      }
-    }
-
-    // зона — только если есть сенсор зоны
-    if (state === consts.STATE_MOWING && isObject(zoneSensor)) {
-      localizedStatus += ` - ${localize('attr.zone')} ${zoneSensor.state}`;
-    }
-
-    // расписание — только если есть next_schedule И party mode выключен (или отсутствует)
-    if (
-      (state === consts.STATE_DOCKED ||
-        state === consts.STATE_IDLE) &&
-      partyMode?.state !== 'on'
-    ) {
-      const nextScheduledStart = this.getEntityByTranslationKey(
-        consts.TK_SENSOR_NEXT_SCHEDULE,
-      );
-      if (isObject(nextScheduledStart)) {
-        const nextDate = new Date(nextScheduledStart.state);
-        if (!isNaN(nextDate.getTime()) && Date.now() < nextDate.getTime()) {
-          localizedStatus += ` - ${this.getEntityName(
-            nextScheduledStart.entity_id,
-          )} ${this.hass.formatEntityState(nextScheduledStart)}`;
-        }
-      }
-    }
-
-    // party mode и lock — только если сущность есть и включена
-    if (partyMode?.state === 'on') {
-      localizedStatus += ` - ${this.getEntityName(partyMode.entity_id)}`;
-    }
-    if (lockMode?.state === 'on') {
-      localizedStatus += ` - ${this.getEntityName(lockMode.entity_id)}`;
-    }
-
-    // ошибка — только если есть сенсор ошибки
-    if (hasError && state !== consts.STATE_RAINDELAY) {
-      localizedStatus += ` - ${this.hass.formatEntityState(errorSensor)}`;
-    }
-
-    return html`
-      <div
-        class="status"
-        @click=${this.handleMore}
-        title=${localizedStatus}
-      >
-        <span class="status-text ${hasError ? 'status-error' : ''}"
-          >${localizedStatus}</span
-        >
-        <ha-circular-progress
-          .indeterminate=${this.requestInProgress}
-          size="small"
-        ></ha-circular-progress>
-      </div>
-    `;
-  }
-
-  /**
-   * Renders a HUI entities card based on the given entities and configuration.
-   *
-   * @param {Array<string>} entities - The entities to be rendered.
-   * @param {boolean} [showCard=false] - Whether to show the card or not.
-   * @return {TemplateResult} The rendered HUI entities card component.
-   */
-  renderEntitiesCard(entities, showCard = false) {
-    if (!this.config || !entities || !showCard) return nothing;
-
-    const entitiesCardConfig = {
-      type: 'entities',
-      entities: entities.map((entity) => ({
-        entity: entity,
-        name: this.getEntityName(entity),
-      })),
-    };
-
-    return this.createHuiCardElement(entitiesCardConfig);
-  }
-
-  /**
-   * Returns a cached or newly created hui-entities-card element.
-   * Re-uses existing elements to avoid DOM thrashing on every render.
-   *
-   * @param {Object} config - The configuration for the HUI entities card.
-   * @return {HuiEntitiesCardElement} The hui-entities-card element.
-   */
-  createHuiCardElement(config) {
-    if (this._huiCardCache.size > 10) {
-      const firstKey = this._huiCardCache.keys().next().value;
-      this._huiCardCache.delete(firstKey);
-    }
-    const key = JSON.stringify(
-      config.entities.map((e) => (typeof e === 'string' ? e : e.entity)),
-    );
-
-    if (this._huiCardCache.has(key)) {
-      const cached = this._huiCardCache.get(key);
-      cached.hass = this.hass; // hass всегда обновляем
-      return cached;
-    }
-
-    const element = document.createElement('hui-entities-card');
-    element.setConfig(config);
-    element.hass = this.hass;
-    this._huiCardCache.set(key, element);
-    return element;
+    return renderCardStatus(this);
   }
 
   async willUpdate(changedProps) {
     if (super.willUpdate) {
       super.willUpdate(changedProps);
     }
-
-    if (changedProps.has('config') || (changedProps.has('hass') && !this._resolvedImage)) {
-      const rawImage = this.config?.image;
-
-      // Резолвим только пользовательские URL; режимам 'ha' и 'default' это не нужно
-      if (this.imageMode !== 'url') {
-        this._resolvedImage = null;
-        return;
-      }
-
-      // Если это специальный URL медиабраузера
-      if (rawImage.startsWith('media-source://')) {
-        // Запрашиваем только если конфиг реально обновился, чтобы не спамить сервер
-        if (changedProps.has('config')) {
-          this._resolvedImage = undefined; // Временно сбрасываем, пока грузится
-          
-          try {
-            // Вызов встроенной функции ядра для получения прямого HTTP-пути
-            const result = await this.hass.callWS({
-              type: 'media_source/resolve_media',
-              media_content_id: rawImage,
-            });
-            this._resolvedImage = result.url; // Получаем путь вида /api/media_source/...
-          } catch (err) {
-            console.warn('Landroid Card: Failed to resolve media image', err);
-          }
-        }
-      } else {
-        // Если это обычная строка (например, /local/my_image.svg)
-        this._resolvedImage = rawImage;
-      }
-    }
+    await this._resolveMediaImage(changedProps);
   }
 
   /**
